+++ apps/hotel/src/preferences/pages/maintenance/maintenance.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { HotelEntity } from '@contler/entity';
-import { MaintenanceModule } from '@contler/models/maintenance-module';
 import { DynamicModuleService, MODULES } from '@contler/dynamic-services';
 import { AuthService } from 'hotel/services/auth.service';
 import { AngularFireDatabase } from '@angular/fire/database';
@@ -16,8 +15,9 @@ import { OptionModule } from '@contler/models';
 })
 export class MaintenanceComponent implements OnInit {
   load = true;
-  modules$: Observable<MaintenanceModule | null>;
+  modules$: Observable<OptionModule[] | null>;
   private hotel: HotelEntity;
+
   constructor(
     private dynamicModule: DynamicModuleService,
     private auth: AuthService,
@@ -27,10 +27,11 @@ export class MaintenanceComponent implements OnInit {
   ngOnInit(): void {
     this.modules$ = this.auth.$employer.pipe(
       tap((data) => (this.hotel = data.hotel)),
-      switchMap((user) => this.dynamicModule.getMaintenanceModule(user.hotel.uid)),
+      switchMap((user) => this.dynamicModule.getOptionsModule(user.hotel.uid, MODULES.maintenance)),
       tap((data) => (this.load = !data)),
     );
   }
+
   public changeStatus($event: MatSlideToggleChange, index: number): void {
     const url = `${MODULES.root}/${this.hotel.uid}/${MODULES.maintenance}/options/${index}`;
     this.db.object<OptionModule>(url).update({ active: $event.checked });
