+++ libs/dynamic-services/src/lib/constants/modules-references.ts
@@ -1,8 +1,8 @@
-export const MODULES = {
-  root: 'modules',
-  immediate: 'immediate',
-  reception: 'reception',
-  room: 'room',
-  maintenance: 'maintenance',
-  form: 'form',
-};
+export enum MODULES {
+  root = 'modules',
+  immediate = 'immediate',
+  reception = 'reception',
+  room = 'room',
+  maintenance = 'maintenance',
+  form = 'form',
+}
