+++ libs/dynamic-services/src/lib/interfaces/form-creation.ts
@@ -1,0 +1,8 @@
+import { InputField } from '@contler/dynamic-services';
+
+export interface FormCreation {
+  form: InputField[];
+  name: string;
+  description: string;
+  icon: string;
+}
